create All Security & fix Post and Comment

+++ src/main/kotlin/com/hw5/hw5/domain/post/service/PostService.kt
@@ -1,41 +1,62 @@
 package com.hw5.hw5.domain.post.service
 
+import com.hw5.hw5.domain.comment.model.toCommentResponse
+import com.hw5.hw5.domain.comment.repository.CommentRepository
+import com.hw5.hw5.domain.exception.ModelNotFoundException
+import com.hw5.hw5.domain.member.repository.MemberRepository
 import com.hw5.hw5.domain.post.dto.CreatePostRequest
 import com.hw5.hw5.domain.post.dto.PostResponse
+import com.hw5.hw5.domain.post.dto.PostResponseWithComment
 import com.hw5.hw5.domain.post.dto.UpdatePostRequest
 import com.hw5.hw5.domain.post.model.Post
 import com.hw5.hw5.domain.post.model.toPostResponse
 import com.hw5.hw5.domain.post.repository.PostRepository
+import com.hw5.hw5.infra.security.MemberPrincipal
 import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 
 @Service
 class PostService(
-    private val postRepository: PostRepository
+    private val postRepository: PostRepository,
+    private val commentRepository: CommentRepository,
+    private val memberRepository: MemberRepository
 ) {
 
     fun getPostList(): List<PostResponse> {
         return postRepository.findAll().map { it.toPostResponse() }
     }
 
-    fun getPost(postId: Long): PostResponse {
-        return postRepository.findByIdOrNull(postId)?.toPostResponse() ?: throw RuntimeException("Post not found")
+    fun getPost(postId: Long): PostResponseWithComment {
+        val post = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        val comment = commentRepository.findByPostId(postId).map { it.toCommentResponse() }
+        return PostResponseWithComment(
+            id = post.id!!,
+            title = post.title,
+            content = post.content,
+            createdAt = post.createdAt,
+            commentList = comment
+        )
     }
 
     @Transactional
-    fun createPost(request: CreatePostRequest): PostResponse {
+    fun createPost(request: CreatePostRequest, memberPrincipal: MemberPrincipal): PostResponse {
+        val member = memberRepository.findByIdOrNull(memberPrincipal.id) ?: throw ModelNotFoundException("Member", memberPrincipal.id)
         return postRepository.save(
             Post(
-               title = request.title,
-                content = request.content
+                title = request.title,
+                content = request.content,
+                member = member
             )
         ).toPostResponse()
     }
 
     @Transactional
-    fun updatePost(postId: Long, request: UpdatePostRequest): PostResponse {
-        val result = postRepository.findByIdOrNull(postId) ?: throw RuntimeException("Post not found")
+    fun updatePost(postId: Long, request: UpdatePostRequest, memberPrincipal: MemberPrincipal): PostResponse {
+        val result = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+
+        if(result.member.id != memberPrincipal.id) throw IllegalArgumentException("회원 ID가 인증된 사용자 ID와 일치하지 않습니다.")
+
 
         result.title = request.title
         result.content = request.content
@@ -43,8 +64,9 @@ class PostService(
     }
 
     @Transactional
-    fun deletePost(postId: Long) {
-        val result = postRepository.findByIdOrNull(postId) ?: throw RuntimeException("Post not found")
+    fun deletePost(postId: Long, memberPrincipal: MemberPrincipal) {
+        val result = postRepository.findByIdOrNull(postId) ?: throw ModelNotFoundException("Post", postId)
+        if(result.member.id != memberPrincipal.id) throw IllegalArgumentException("권한이 없는 사용자입니다.")
         return postRepository.delete(result)
     }
 }
